requirements.txt
amqp==2.6.1
asgiref==3.2.10
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
cffi==1.14.1
channels==2.4.0
chardet==3.0.4
click==7.1.2
constantly==15.1.0
coverage==5.2.1
cryptography==3.0
daphne==2.5.0
dj-database-url==0.5.0
Django==3.0.8
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-cors-headers==3.4.0
django-extensions==2.2.9
django-filter==2.3.0
django-timezone-field==4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
h11==0.9.0
httptools==0.1.1
hyperlink==19.0.0
idna==2.10
incremental==17.5.0
kombu==4.6.11
Pillow==7.0.0
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.10
uvicorn==0.11.8
uvloop==0.14.0
vine==1.3.0
websockets==8.1
zope.interface==5.1.0
